requirements.txt
pymediainfo
PyQt6
Pillow
pillow-heif
